requirements.txt
streamlit
pytesseract
pillow
pandas
fpdf
